Promises
In the api.js we removed the callback if statement. we then created a new promise where we kept the parameters of the previous function. Then we made a statement to handle the err and the success. We then followed all the callbacks and fixed them to make it compatible with a promise.